[kernel] Replace uint32_t with more relevant types

+++ kernel/include/algo/bitmap.h
@@ -13,12 +13,12 @@
 
 struct bitmap {
     uint8_t* data;
-    uint32_t len;
+    size_t len;
 };
 typedef struct bitmap bitmap_t;
 
-bitmap_t bitmap_wrap(uint8_t* data, uint32_t len);
-bitmap_t bitmap_allocate(uint32_t len);
+bitmap_t bitmap_wrap(uint8_t* data, size_t len);
+bitmap_t bitmap_allocate(size_t len);
 int bitmap_find_space(bitmap_t bitmap, int req);
 int bitmap_find_space_aligned(bitmap_t bitmap, int req, int alignment);
 int bitmap_set(bitmap_t bitmap, int where);
