+++ kernel/include/platform/aarch32/tasking/context.h
@@ -17,12 +17,12 @@ typedef struct {
     uint32_t lr;
 } PACKED context_t;
 
-static inline uint32_t context_get_instruction_pointer(context_t* ctx)
+static inline uintptr_t context_get_instruction_pointer(context_t* ctx)
 {
     return ctx->lr;
 }
 
-static inline void context_set_instruction_pointer(context_t* ctx, uint32_t ip)
+static inline void context_set_instruction_pointer(context_t* ctx, uintptr_t ip)
 {
     ctx->lr = ip;
 }
