+++ kernel/include/platform/x86/tasking/context.h
@@ -21,12 +21,12 @@ struct PACKED context {
 };
 typedef struct context context_t;
 
-static inline uint32_t context_get_instruction_pointer(context_t* ctx)
+static inline uintptr_t context_get_instruction_pointer(context_t* ctx)
 {
     return ctx->eip;
 }
 
-static inline void context_set_instruction_pointer(context_t* ctx, uint32_t ip)
+static inline void context_set_instruction_pointer(context_t* ctx, uintptr_t ip)
 {
     ctx->eip = ip;
 }
